+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes to this project will be documented in this file.
 
 - move login form hook create log event call into server sign in function
 - move register form hook create log event call into server sign up function
+- move lost password form hook create log event call into server lost password function
 
 #### Services
 
@@ -28,6 +29,11 @@ All notable changes to this project will be documented in this file.
 - log event edge function to accept source value
 
 ### Removed
+
+#### Schemas
+
+- login form return schema
+- register form return schema
 
 #### Server